Safeguard against empty item object :bug: -- :closes: #6892

+++ inc/Engine/Common/PerformanceHints/Database/Queries/AbstractQueries.php
@@ -60,7 +60,11 @@ class AbstractQueries extends Query {
 
 		$deleted = true;
 		foreach ( $items as $item ) {
-			$deleted = $deleted && $this->delete_item( $item->id );
+            if( is_object( $item )  && isset( $item->id ) ) {
+                $deleted = $deleted && $this->delete_item( $item->id );
+            } else {
+                $deleted = false;
+            }
 		}
 
 		return $deleted;
